# Improve scalar string replace performance for long strings