Fixes #7370.

This adds a scalar string replace algorithm with character-level parallelism which significantly improves the performance of scalar string replacement on longer strings.  It can involve many more kernel launches than the row-based algorithm and does not always outperform on short strings.  Therefore a heuristic based on the average character length of valid string rows is used to automatically select which algorithm to use.