@jlowe 2021-02-19
Performance comparisons from the scalar benchmark.
Before:
```
---------------------------------------------------------------------------------------------------------------------
Benchmark                                                           Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------------------------------
StringReplaceScalar/replace_scalar/4096/32/manual_time          0.151 ms        0.168 ms         4367 bytes_per_second=459.181M/s
StringReplaceScalar/replace_scalar/4096/128/manual_time         0.422 ms        0.440 ms         1653 bytes_per_second=643.174M/s
StringReplaceScalar/replace_scalar/4096/512/manual_time          2.48 ms         2.49 ms          282 bytes_per_second=442.666M/s
StringReplaceScalar/replace_scalar/4096/2048/manual_time         27.3 ms         27.3 ms           26 bytes_per_second=159.837M/s
StringReplaceScalar/replace_scalar/4096/8192/manual_time          453 ms          453 ms            2 bytes_per_second=38.304M/s
StringReplace/replace_scalar/32768/32/manual_time         0.166 ms        0.184 ms         3998 bytes_per_second=3.24747G/s
StringReplaceScalar/replace_scalar/32768/128/manual_time        0.452 ms        0.471 ms         1545 bytes_per_second=4.71488G/s
StringReplaceScalar/replace_scalar/32768/512/manual_time         2.71 ms         2.73 ms          258 bytes_per_second=3.13709G/s
StringReplaceScalar/replace_scalar/32768/2048/manual_time        34.6 ms         34.7 ms           20 bytes_per_second=1009.18M/s
StringReplaceScalar/replace_scalar/32768/8192/manual_time         526 ms          526 ms            1 bytes_per_second=265.082M/s
StringReplaceScalar/replace_scalar/262144/32/manual_time        0.337 ms        0.356 ms         2068 bytes_per_second=12.8219G/s
StringReplaceScalar/replace_scalar/262144/128/manual_time        1.55 ms         1.57 ms          448 bytes_per_second=10.9859G/s
StringReplaceScalar/replace_scalar/262144/512/manual_time        17.0 ms         17.1 ms           41 bytes_per_second=3.99909G/s
StringReplaceScalar/replace_scalar/262144/2048/manual_time        193 ms          193 ms            4 bytes_per_second=1.41648G/s
StringReplaceScalar/replace_scalar/2097152/32/manual_time        1.84 ms         1.86 ms          379 bytes_per_second=18.7388G/s
StringReplaceScalar/replace_scalar/2097152/128/manual_time       9.47 ms         9.49 ms           74 bytes_per_second=14.4086G/s
StringReplaceScalar/replace_scalar/2097152/512/manual_time       88.8 ms         88.8 ms            8 bytes_per_second=6.14807G/s
StringReplaceScalar/replace_scalar/16777216/32/manual_time       13.9 ms         13.9 ms           50 bytes_per_second=19.9076G/s
```
After:
```
-----------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                               Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------------------------------------------------
StringReplaceScalar/replace_scalar_autoalg_single/4096/32/manual_time               0.152 ms        0.169 ms         4495 bytes_per_second=456.924M/s
StringReplaceScalar/replace_scalar_autoalg_single/4096/128/manual_time              0.113 ms        0.131 ms         5786 bytes_per_second=2.33732G/s
StringReplaceScalar/replace_scalar_autoalg_single/4096/512/manual_time              0.150 ms        0.166 ms         4419 bytes_per_second=7.1418G/s
StringReplaceScalar/replace_scalar_autoalg_single/4096/2048/manual_time             0.301 ms        0.318 ms         2319 bytes_per_second=14.1505G/s
StrilaceScalar/replace_scalar_autoalg_single/4096/8192/manual_time             0.946 ms        0.964 ms          730 bytes_per_second=17.9302G/s
StringReplaceScalar/replace_scalar_autoalg_single/32768/32/manual_time              0.165 ms        0.183 ms         4139 bytes_per_second=3.26231G/s
StringReplaceScalar/replace_scalar_autoalg_single/32768/128/manual_time             0.198 ms        0.214 ms         3403 bytes_per_second=10.7608G/s
StringReplaceScalar/replace_scalar_autoalg_single/32768/512/manual_time             0.513 ms        0.532 ms         1288 bytes_per_second=16.5836G/s
StringReplaceScalar/replace_scalar_autoalg_single/32768/2048/manual_time             1.86 ms         1.87 ms          376 bytes_per_second=18.3948G/s
StringReplaceScalar/replace_scalar_autoalg_single/32768/8192/manual_time             7.57 ms         7.59 ms           92 bytes_per_second=17.9735G/s
StringReplaceScalar/replace_scalar_autoalg_single/262144/32/manual_time             0.336 ms        0.355 ms         2061 bytes_per_second=12.8375G/s
StringReplaceScalar/replace_scalar_autoalg_single/262144/128/manual_time            0.952 ms        0.970 ms          719 bytes_per_second=17.9219G/s
StringReplaceScalar/replace_scalar_autoalg_single/262144/512/manual_time             3.73 ms         3.75 ms          188 bytes_per_second=18.29G/s
StringReplaceScalar/replace_scalar_autoalg_single/262144/2048/manual_time            15.6 ms         15.6 ms           45 bytes_per_second=17.5188G/s
StringReplaceScalar/replace_scalar_autoalg_single/2097152/32/manual_time             1.84 ms         1.86 ms          379 bytes_per_second=18.7479G/s
StringReplaceScalar/replace_scalar_autoalg_single/2097152/128/manual_time            7.68 ms         7.69 ms           91 bytes_per_second=17.7849G/s
StringReplaceScalar/replace_scalar_autoalg_single/2097152/512/manual_time            32.0 ms         32.0 ms           22 bytes_per_second=17.0394G/s
StringReplaceScalar/replace_scalar_autoalg_single/16777216/32/manual_time            13.9 ms         13.9 ms           50 bytes_per_second=19.9068G/s
```